Created a mechanism for user profiles. For now, this just allows users to have a personal homepage URL, along with a flag indicating their whitelisted status. Users can edit their own profiles, but cannot change their own whitelist status. There's an admin interface included so the site administrator can whitelist certain users.
At Eric H's suggestion, the model was added to the core app; there's a South migration to create the new table (./manage.py migrate core), and a new dependency (pip install django-profiles).

User profiles can be viewed at /profiles, and your own profile can be created at /profiles/create, then edited at /profiles/edit. The detail of any user's profile can be viewed at /profiles/<username>. We may want to eventually tie this in with a list of that user's projects.

There are currently no visible links to any of the profile stuff; it might be good to at least add a link from the dashboard to create your profile. It may also make sense to automatically create an empty profile for any new users that sign up. None of this stuff is addressed in this initial version.