Added language support to our urls/subdomains.
This is mainly a no-op on the backend, and they are all hard-coded to 'en' for now. This is the only thing that pretty much makes it possible to do this (otherwise we'd have to do some fun logic to check if something was a version, or a valid language code).

Because documentation paths can look like "/docs/django/howto/testing.html", we have to account for howto matching a version, and validate that it is indeed. If it isn't a valid version, we have to append them back together and pass them along. And lots of other fun bits like that. So I decided to future proof the URLs as much as possible.