Support a Standard RTD Theme!
This only applies the RTD theme to projects that are using 'sphinxdoc' or 'default' themes. So if you have your own, currently you will continue to see it.

Also vastly simplified the whitelisting stuff, making the whitelisted and non-whitelisted logic the same. This takes advantage of conf.py being python, and just embeds some logic at the end of the conf.py file for all projects to inject RTD stuff.