/page/ URL redirects to page in default version
As I suggested in #282, this adds a URL pattern that can redirect to a specific
page under the default version. Using "/page/..." either on readthedocs.org or
a subdomain redirects.

Note that the redirect goes to "/docs/<project_slug>/..." even on subdomains.
This could be fixed by using a different URL pattern name for the subdomain
URLs and specifically requesting that in the call to `reverse()`.