Move GitHub links to "This Page" section
IMHO the GitHub links had a too prominent location in the sidebar, prioritized
higher than the ToC for the current page and the docs navigation. This patch
moves the links down to the "This Page" section in the bottom of the sidebar.
The "This Page" section is also a natural location for the "Show/Edit on
GitHub" links as it already contains the similar "Show Source" link.