Fix typo in notification
While waiting for docs to be built, users receive a notication that should read
'Your docs,' not 'You docs.' The typo is being fixed in both the template and
the translation files.

It appears that there was also some trailing whitespace in the translation
files. Trimming it should be a trivial change.