Remove the ability to upload HTML.
This was a half supported feature,
and goes against the general idea of Read the Docs.
Having the Repo <-> HTML mapping is a fundamental part of Read the Docs,
and breaking that expection of the user is bad.

In an effort to remove code, and embrace constraints,
this is functionality is deprecated,
and will be removed in a future commit.