Change the builder so that the version sync code is on the server.
This will make builds much faster, and reduce a lot of overhead around HTTP calls. It should keep the versions in the repo and the versions on the server properly in sync. This has been a long standing bug where people have changed tags and we wouldn't update the code on checkout correctly.