Fix activation email typo.
It looks like there used to be some pluralization logic once upon a time, but
it got removed (presumably for a good reason).

I've no idea where `expiration_days` gets set (it doesn't seem to occur
anywhere other than this file), but on the actual site right now it's set to 7.
And 7 is plural, so saying "7 day" is wrong and awkward.  Now it's not wrong or
awkward.