RFC: Clear VCS checkout before cloning (closes #858)
Please note: this commit hasn't actually been tested.

If a project repo type is changed, RTD fails to check it out.

This is probably because VCS tools generally fail to clone a new repo in
a directory which already has files.

This commit will delete all files in the VCS working dir before cloning
a new repo.