Place `projects` before `builds` in INSTALLED_APPS
When attempting to install readthedocs from scratch, installation fails because the migrations in `builds` depend on migrations from `projects`.  While these dependencies can be explicitly modelled in newer versions of South, they are implicitly modelled in older versions via app ordering.

This commit fixes the installation, but a better solution might be to update to the latest South and model this dependency explicitly.