Add a one click import of demo project
This adds a view that performs a mocked form submission to create demo project.
It will create a new project with the name `<username>-demo` using our template
repository. If a project exists with the same repo, don't reimport. If a project
exists with the same name, throw an error and drop back to the dashboard.