Only use the current_page if it is available in the MkDocs template
As of the development version of MkDocs, the template is used to build a
few extra templates like the 404.html page. While these may not be used
on ReadTheDocs, they will still be built.

On these pages, the context is very similar, but there is no current
page as that object represents a Markdown file.