Removing use of User.get_profile and settings.AUTH_PROFILE_MODULE.
Django 1.8 does no longer support User.get_profile and switching out the
profile model. We should change this in the long run to use a swappable User
model instead. For now we hardwire the core.UserProfile as profile model and
the core.UserProfileForm as profile form wherever it is used.