Delete pdf/epub artifacts when their build is disabled
The reason for this is that old versions which had PDF builds in the past and
for which the PDF builds have been disabled will keep the stale PDF of old
builds lying around. We need to delete the artifacts if they exist and are
disabled for the current build.

Addresses https://github.com/rtfd/readthedocs.org/pull/1431#issuecomment-120998189