Replace Version.get_vcs_slug() and Version.remote_slug with Version.commit_name
The two places were solving the same problem but with different results. That
is undesired and resulted in invalid behaviour. The commit_name property
unifies those now and has tests to make sure it works the way intended.

The problem with this on-first-sight-easy problem is that we currently do
store different things in "identifier", "slug", "verbose_name" fields
depending if its a branch or tag, if it's a special version (like
stable, latest) etc. The cleanest solution would be to refactor the Version
model to expose those information more easily and store branch_name, tag_name,
explicitly.