Changes to support Bitbucket OAuth2
Changes most API calls to use APIv2 and relies on OAuth2 for sessions. Bitbucket's OAuth2
implementation expires tokens after 2 hours, so this also implements token refreshing on
the general session creation.

For now, this relies on a patched version of django-allauth. A new ticket should be created
to track migrating django-allauth back to upstream once the PRs are merged in and released.

Refs pennersr/django-allauth#1215
Refs pennersr/django-allauth#1212