Fix several issues with build environments
This fixes #1818, which is a bug triggered by the combination of long project
and/or version names. The container hostname was longer than the default 64
characters allowed in Ubuntu.

A new container_id is used, which omits the version name for just the build id,
including also the project id, followed by the project name. All of this is
truncated to the max hostname limit. The project id and name are only for
operations purposes.

This also fixes #1978, where the original pass on builds would briefly label the
build as successful. An option for skipping successful build notifications was
added to the base `BuildEnvironment`