Add a per-project advertising opt-out, and form page for preference
This adds a per-project advertising opt-out form, with some guidance on supporting
Read the Docs if you oppose advertising. It links to several pages explaining more
about the advertising spot, how to opt-out of ads, and a backup option to just
remove ads altogether if you don't want to support the project.