Add session configuration to API clients
This also cleans up the api v1 client to match the implementation by api v2.
A session is added to the Slumber API to handle an explicit hostname on the
request. This is useful for local development and staging, where the API
host name might not match the host name the web server or application instance
is designated to listen for. For example, when requests for API should go to
http://localhost, but the application expects http://readthedocs.org