Add a mixin class for dashboard views on models with project relations
As we overhaul the project admin dashboard, a good amount of code can be cleaned
up by using a CBV rather than repeating the view code like is currently used. I
had a need for this outside this code base, but implemented it here instead.

Views would look like:

    class DomainMixin(ProjectRelationMixin):
        model = Domain
        lookup_url_kwarg = 'pk'

    class ListDomainView(DomainMixin, ListView):
        pass

    class DetailDomainView(DomainMixin, DetailView):
        pass

Views would then have access to a `project` and `domains` in template context
data.