SingleHtmlBuilder; impl __init__ to fix singlehtml
In a previous commit, an __init__ method was added to HtmlBuilder and to
HtmlDirBuilder, however the SingleHtmlBuilder was unmodified. Since
SingleHtmlBuilder inherits from HtmlBuilder, HtmlBuilder.init is
executed which sets the instance variable "sphinx_builder" to either
"readthedocs" or "readthedocs-comments".

This instance variable from HtmlBuilder makes it impossible to get the
correct value from SingleHtmlBuilder's class variable "shpinx_builder"
"readthedocssinglehtml".

The fix is to implement the __init__ method for SingleHtmlBuilder and
set self.sphinx_builder accurately.