Remove submodule docs/_themes/sphinx_rtd_theme
The submodule was only relevant for a short time and its continued
existance is a distraction for new contributors.

See: https://github.com/snide/sphinx_rtd_theme/pull/342

Using git-blame, I was able to pinpoint which commite introduced the
submodule, but I wanted to be certain that it was no loger in use. So, I
manually inspected parts of the source code, ran `git grep ...`,
reviewed open and closed issues, and PRs, but found nothing conclusive
about its existance. Ultimately, git-log is what gave me the most
confidence that it is no longer necessary. As you can see, it was
introduced in October of 2013 and then made obsolete a month later.

From git-log:
  bash$ git log -G '_themes/sphinx_rtd_theme'
  commit 50619d1bc514c99bcada50a7daf4f7ac71e7205e
  Author: Eric Holscher <eric@ericholscher.com>
  Date:   Fri Nov 8 18:55:06 2013 -0800

      Rejigger theme stuff in our own conf

  commit d9829a2d815193fbbe80cb3d507a2948b56c3794
  Author: Eric Holscher <eric@ericholscher.com>
  Date:   Thu Oct 17 13:26:19 2013 -0700

      Add new theme