Fix broken celery test
This is really a design decision thing. Either projects should be marked
as having valid clones even if record is off, or they shouldn't.

Personally, I believe the former behavior is better, since we will have
actually cloned the repo at that point (in the setup_env), but this does
mean that record will record *some* data in the database, which might
not be desirable.