Initial rollout of Sentry integration with 404/500’s
This adds a new Error Page promo type,
and adds the data into the error page rendering.
It also adds the Sentry “User Feedback” UI to our 500 pages.
I’m not 100% sold on this UI,
but it’s pretty neat to get user feedback.
I wish it was a bit more configurable. 

This isn’t using the JS library currently,
so it is hard-coding the promo text.
I should probably enable injecting dynamic copy for this over time,
but this seems fine for now.