Add ad support for Alabaster theme
This updates our ad insertion code to also support the Alabaster theme. The ad
is injected currently at the bottom of the sidebar, below the search box. Some
of styling was updated for this as well.

This also makes a minor change to the CSS. `div.rst-pro` was inexplicably named,
perhaps to match something in the sphinx_rtd_theme css/html? I changed this to
`div.rtd-pro`, which makes more sense. The old rules were kept in the odd case
that the old classes/promo html were used/cached somehow -- this shouldn't be
the case though.

Also, some of the dependencies were updated.