Split index_search_request into several functions.
By splitting the function up, we not only resolve pyflakes issues around local
variables, we also make this function easier to test.

I don't like the SectionIndexer approach, but there was a chunk of state
(section_index_list) which was extended in each loop iteration. To recreate
that functionality while limiting the number of lines in each loop, the state
is moved into a separate object. If the original logic was faulty, this could
be much more legible.