Explicitly wrap forms paramter from django-formtools in a list.
django-formtools is passing us the `items()` of a `dict`. In Python 2,
this is a `list`, and can be directly indexed. In Python 3, it is a
"View Object", which cannot be directly indexed.

Since I thought it unlikely for there to be more than a few forms in this
collection, I decided it was easier to simply wrap it in a `list` no
matter what, rather than using `itertools.islice` to grab the first
element.