New subproject admin page (#2957)
* Remove privacy app

The privacy app was a strange mixture of various application models managers,
querysets, and syncers. Instead, logic is moved to where we should be using it,
inside the other applications

* New subproject admin page

This replaces the list of text with a table that is more navigable and also
replaces the create form with a dropdown form of the projects you can add. This
list of projects is limited to projects you are the admin of, and which have not
been added as a sub or super project anywhere else.

This depends on #2954 and others.

* Lint fixes

* Require arguments to project form