Intro Dockerfile to ease help adoption.
Users of this Dockerfile can be either contributors or administrators.

For contributors, they will be able to make changes and be able to see
their changes by rebuilding the image.

For administrators, they will be able to quickly evaluate RTD without
having to contribute a lot of time.

Build:
$ docker build --tag rtfd-server .

Run:
$ docker run -p 8000:8000 rtfd-server