Fix issue with CORS headers on subproject projects
We weren't correctly setting CORS headers when querying the documentation search
API from a subproject on and parent project domain. In this case, we were
querying with a project lookup for the subproject, and no domain was assigned to
this project.

This alters the query to find a domain associated with either the project you
are querying, or the parent project of a project you are querying.