Do not update version type when sync versions
This allows us to have the same `verbose_name` for different type of
versions (for example, '2.0' could both a tag and a branch). If this
happen, we want to keep them both.

In the UI, it will say just `"2.0"` for the branch and
`"2.0 (c0d25453)"` for the tag. This could be improved later if
necessary.