Change stable version switching to respect `active`
This previously used the `machine` tag that was turned off when a user de-activated the version,
but never turned back on when they re-activated it.

This changes to just use the `active` flag on the Version,
so we will update the stable version if the existing one is `active`,
and not if it isn't,
which makes _way_ more sense.