Update style and copy on abandonment docs
I went to look up our new policy and got confused :)

"Maintainers" is a term I attribute to project maintainers, not the core team of
Read the Docs. Because we refer to "core team" regularly, I updated that here.

This also alters the style of the lists to be less formal legal structure and
adds a couple pieces of reST that fit better.