Add concept of dashboard analytics code
This lets us have a specific Analytics code for the dashboard,
instead of using it for dashboard & docs.
This keeps the pre-existing global one for all things,
but lets us segment out the dashboard.

This is mainly useful for having GA not sample our dashboard traffic,
so we can get real data for it.