Add view to detail a project with its subprojects
We need the concept of an organization that will group more than
one projects. To do that we can leverage the project / subprojects
relationship by just adding a view.
The style is the one of the homepage as this must only be a page
between the homepage and the single project detail view.