Update instance and model when `record_as_success`
When we record the command as SUCCESS (exit_code=0) we also need to
update the current instance of BuildCommand because `self.exit_code`
is used to calculate BuildEnvironment.successful. The later is part of
the decision of `update_on_success`, and as the command was marked as
failed, we were updating the status.

This looked as a random failure, but it was happening always that any
of the `record_as_success` command failed, but it wasn't easy to see
it because it the jQuery request has to be done exactly after the
Build was marked as FAILED and before the next command was executed,
since the state was changed immediately to "Building"