This fixes two bugs apparent in nesting of translations in subprojects
The canonical domain was not searching as deep as we support. We support
translations nested in subprojects, and vice versa, this was previously only
searching for translation or subproject, so no nesting at all.

The path was also a similar problem, but instead of a heavy refactor that reused
the canonical domain lookup, I just altered the logic a little here. The pattern
was required to support both superproject as a translation project and
subproject as a translation project.