Don't share state in build task
Move task steps to an isolated class, instead of possibly sharing state in
between tasks. This bug is not consistently reproduceable, but one consistent
way to reproduce is to drop to ``rdb`` on the UpdateDocsTask run method.
Previously, if you did this while running 2 celery processes, and triggered 2
builds, you could witness state changing on the task instance. By moving state
to a separate class that is instantiated on each task execution, this completely
avoids shared state in these tasks.