Remove django-celery-beat and use the default scheduler
There are some issues with timezone for the `DatabaseScheduler`, so we
fallback to the default one (celery.beat.PersistentScheduler) which is
simpler --doesn't allow us to change the scheduled task from the
admin, but we don't need that.

In case we need this again, we can come back when those issues are
already fixed and revert this commit.