Give TaskStep class knowledge of the underlying task (#3983)
There is no `self.retry` on the task step class, this exists on the task.

This also fixes a bug where we weren't handling the lock on the python step as
well.

We discussed a better pattern to use here would be to rely on celery on_failure
and on_retry event methods more. This would simplify logic here greatly.