Override versions.html template for specific RTD functionality
Sphinx RTD theme uses a `versions.html` template to generate the
version menu if `READTHEDOCS` variable is True. Since this is
something specific to Read the Docs and that variable is going to be
removed soon from the theme, we are overriding this template to
maintain our current functionality.

Also, the static content generated in this menu is removed to avoid
showing invalid links/data in case the AJAX fails for any reason. If
this happen, the menu just won't show any information.