Use PATHs to call clear_artifacts
Use a list of PATHs Instead of using a ``Version.pk`` when calling
`clear_artifacts`. This is because in the ``Version.delete`` method,
the Version object is removed immediately after broadcasting the
``clear_artifacts`` task and that could produce a race condition.

With this change, we calculate all the PATHs needed _before_
broadcasting the task and _before_ deleting the object from the
database.