Depend on reliable data from the db for these tests
This Feature is created inside a db migration. As we depend strictly
on this Feature we can't rely on data from the migration since it's
ran only once when the db is created and not re-populated on each
test.

We need to be sure that the data we rely on is in the database in a
valid state and that it wasn't modified/deleted by any other test
case. Because of this, we need to use a fixture or delete and create
the data we rely on inside the ``setUp`` for this test.