Cache the footer_api template based on version for one minute
This is something that should increase the speed of our responses on the most popular versions.
It will make it so that users who update their versions might not see the footer change in under a minute,
but I think we can probably put up with that for the hopeful performance benefits.