Use Docker time limit for max lock age
When building a project, if it tooks more than `REPO_LOCK_SECONDS` and
while building after that time another build is triggered for the same
Version and the same builder takes the task the lock will be
considered "old" and remove and taken by the new build.

This will end up in a collision when accessing the files and it could
raise an exception like `IOError: [Errno 26] Text file busy`. Also, it
could fail with another unexpected reasons.

This PR increases the `max_lock_age` to the same value assigned for
the project to end the build in order:

* custom container time limit or,
* `settings.DOCKER_LIMITS['time']` or,
* `settings.REPO_LOCK_SECONDS` or,
* 30 seconds

Related to #1609