Feature flag to make `readthedocs` theme default on MkDocs docs
Historically, we were using `readthedocs` as default theme for MkDocs
but in https://github.com/rtfd/readthedocs.org/pull/4556 we decided to
change it to get the same behavior when building locally than in Read
the Docs.

This commit adds a Feature flag to keep having the old behavior for
some particular projects so we can add these project and do not break
their documentation (change the theme without asking/reason).