Remove `project_version_delete_html` view
We used to have a "Clean" button in the Inactive Versions listing to
remove built and disabled versions of the documentation.

This is not more possible, because when the version is disabled (mark
as non active) we call `clean_artifacts` and remove all the HTML files
for this version and mark it as `Version.build=False`. In the end,
that button does not appear anymore.

Related #4937