Use django-textclassifier to validate Project.name and description
This classifier gives use scores for VALID or SPAM of the name and
description of the project that it's about to be imported.

If the accuracy for SPAM is greater than 0.95 we block the import.

This commit also adds useful admin action to mark project as VALID or
SPAM (optionally delete them) to increase the accuracy when checking
new imports.