Remove header nav (Login/Logout button) on 404 pages
We are using the same 404.html template for 404 errors when serving
documentation (.io) and also on our platform (.org).

The Login/Logout button only shows the proper word when the 404 page
comes from .org, but it always says "Login" when we hit .io 404 pages.

As this is only confusing and there is not value added here, I'm
removing it completely.