Add to and reorder GitHub webhook docs
This does a small cleanup on the GitHub webhook docs by reordering the instructions in the order which fields are shown to the user on GitHub (reducing jumping around) but also explains the initial "ping" process that GitHub does.  It was confusing to see RTD respond with `{"detail":"Unhandled webhook event"}` to a ping, when in fact this is working-as-expected.