Use downloads.html in template
In #5502 we refactored the function
but we used to depend on `pretty` to change the name of the key from
`html` to `htmlzip`, we use `html` in the footer and in the theme.

So, it's safe to change the value in the template.

See https://github.com/rtfd/readthedocs.org/pull/5502/files#r272787188