Hide "Protected" privacy level from users
Protected causes a lot of confusions to users. I'm hiding it from the
Form for now as a temporarily solution while we implement more Version
states and this behavior can be easily managed.

See https://github.com/rtfd/readthedocs.org/issues/5321

Project/Version that are Protected will remain in the same state and
everything should keep working. Although, new Project/Version are not
allowed to set as Protected.