Refactor redirect tests
This is most like a little hack to make
the tests of the new feature in .com pass.

Why tests are passing after this changes?

First, we don't need to patch anything, by default
those files don't exist.

Now, here is the interesting part.
When a version exists (in this case `0.8.1`)
the serve_docs function (from .com)
tries to verify that the current user has permissions over that
version first, this gives a 401 response.

Accessing to a unexisting version, it gives 404,
which is needed to trigger the redirect