Clean up permissions
We never assign this permissions
inside the code, the only way I can see them
is from the shell.

Also, I don't think this is necessary,
we have the `is_superuser` check to have special access.
But I think maybe the admin is a better fit for things like this?