Update conda at startup
Behind a Feature flag, we can decide if we want to update the version
of ``conda`` from the system to its latest version available.

This allow us to use a new release of ``conda`` that is not released
under Miniconda yet, but also to not upgrade the Docker image just
because of a new release of Miniconda.

(the overhead added is around 60 seconds in total)