Update get_absolute_url for External Versions
This allows us to properly set a new domain in production for the external version hosting.
It's a bit hacky for now,
mostly because we don't have a way to specify HTML or PDF versions of the external versions.

We will need to think more about the UX around multiple types,
but that's also an existing issue with our current doc builds (we only link to HTML).