Don't rely on instance to save object
There is not reason to rely on the instance
(well, just save one query).

We are having a weird bug on the corporate site
I wasn't able to replicate it locally
or find any other explanation.

Hope that fetching the object in an explicit way
fix the bug.

Also, we could just rely on the object id instead
of the slug, that way there is not need to fetch
the object in the save method :)