Return full path URL (including .html) on /api/v2/docurl/ endpoint
`make_document_url` had a  behavior and it was returning invalid URL
when the `page` argument was a docname (the case when used by the
Embed API endpoint).

This commit consider both cases for this attribute (filename page:
index.html, or document name: config/v2) and decide if the `.html`
extension needs to be added based on the builder
--`documentation_type` (sphinx, sphinx_htmldir, mkdocs, etc) used for
the project.