Improve CORS configuration
The goal of this commit is to allow access to our resources from
different hosts:

* everything coming from `*.readthedocs.io`
* everything hitting any URL under `/api/` if it's safe method, no
matter the host where the request comes from

Compared with the previous implementation, this reduces the number of
queries to our database since we are not checking that the host where
the request was made is a known Domain for us.

Besides, this opens access to `/api/v3` as well.