Alter settings layout to match commercial
In trying to get this working for commercial, it became apparent that
the wrong settings layout was used for this. Instead of having a
dev.celery.* settings branch, the descending importance naming scheme
used in readthedocsinc was used instead.

That is, the settings file should be called
`readthedocs.settings.build.dev`, not
`readthedocs.settings.dev.celery.build`.