Do not register `/_/downloads/` URL in web instance
Remove the URL for `/_/downloads/` from the web instance and only
register it in El Proxito who is whom is going to serve these files.

Since we don't have this URL in the web instance, we need to generate
it manually (we can't user `reverse`) from the web instance when
returning the production media URL of a Project/Version.