Execute checkout step respecting docker setting
Closes #5673

There is a couple of things with this:

- Git commands using gitpython will use the git binary from the host
- Git commands using git directly, will use the git binary from the
  container.

This may bring some surprises, since the git versions will be different.
But I'm not really worried about it, it should work.

Another thing is: To make a clone we will spin a build image, which are
big, probably we could add an image that only has git in it.

This should be merged after we have integrated this with the ssh-agent
in .com