Improve Proxito 404 handler to render user-facing Maze when needed
Use a specific 404 handler in El Proxito to decide if it's called to
be executed as a NGINX internal redirect (`fast_404` view) or if was
called to render the 404 user-facing page (`ServeError404` view).

This way, we keep the fast 404 view for all our internal handling but
we use the standard Django flow to render 404 pages for user.