Do not mention site-packages anymore
We don't really support these libraries. It's not possible to know
which versions are installed and, the libraries currently installed in
the images by default are not a problem in our builds.

If we want to go in this direction, we should add problematic
libraries (like `torch`, `tensorflow`, etc) and document what versions
are in the docker images.