Ignore failing redirects (from_url smaller than 5 chars long)
We first perform a `.count()` query to check if the DB SQL query will
fail. If it fails, we ignore the `exact` Q object and we continue with
the redirects check ignoring all the Exact Redirects for this project.

This query will *only fail* if there is at least one Exact Redirect
with less than 5 chars long on this project. This is because we are
substracting the `$rest` part from the `from_url` field and we can't
end up with a negative number.