Make search compatible with sphinx 2.2.1
We are relying on FILE_SUFFIX,
but we were relying on a bug actually https://github.com/sphinx-doc/sphinx/pull/6741

This is compatible with older and new versions.

This is also related that we don't save the real full path when
indexing.

https://github.com/readthedocs/readthedocs.org/blob/24ca737e3f5c0df359801a8dede6519f67a70cb6/readthedocs/search/api.py#L34-L38

Fixes https://github.com/readthedocs/readthedocs.org/issues/6801